python-metrics-app/app/requirements.txt
Flask==2.3.2
psycopg2-binary
prometheus_client
